use int8_t

+++ csrc/utils.cpp
@@ -280,8 +280,8 @@ bool is_cpu_scalar(const c10::TensorType& tensor_type) {
 // Check device of TensorType in all inputs ensure all tensors are on cuda
 // devices.
 // return common device index (or -1 if device differs).
-int getCommonDeviceCUDA(const at::ArrayRef<c10::IValue>& inputs) {
-  int index = -1;
+int8_t getCommonDeviceCUDA(const at::ArrayRef<c10::IValue>& inputs) {
+  int8_t index = -1;
   size_t num_tensors = 0;
   for (const auto& input : inputs) {
     if (!input.isTensor()) {
@@ -297,7 +297,7 @@ int getCommonDeviceCUDA(const at::ArrayRef<c10::IValue>& inputs) {
     if (index != -1 && index != cur_index) {
       return -1;
     }
-    index = (int)cur_index; // NOLINT
+    index = cur_index;
     ++num_tensors;
   }
   // A case where there is only a scalar input should not indicate a failure
